edit

+++ app/Http/Controllers/Admin/ReceiptBranchController.php
@@ -149,6 +149,9 @@ class ReceiptBranchController extends Controller
     {
         $receipt_branch_product_pivot = ReceiptBranchProductPivot::find($id);
         $receipt = ReceiptBranch::find($receipt_branch_product_pivot->receipt_branch_id); 
+        if($receipt->done){
+            return 0;
+        }
 
         $receipt_branch_product_pivot->delete();
 
@@ -177,6 +180,10 @@ class ReceiptBranchController extends Controller
 
             $receipt_product_pivot = ReceiptBranchProductPivot::find($request->receipt_product_pivot_id);
             $receipt = ReceiptBranch::find($receipt_product_pivot->receipt_branch_id);
+            if($receipt->done){
+                toast('لا يمكن تعديل المنتجات بعد الإنتهاء من الفاتورة','error');
+                return redirect()->back();
+            }
             $price_type = $receipt->price_type();
             
             $product = ReceiptBranchProduct::findOrFail($request->product_id); 
@@ -216,6 +223,10 @@ class ReceiptBranchController extends Controller
             return view('admin.receiptBranches.partials.add_product',compact('products','receipt_id','order_num','price_type'));
         }else{
             $receipt = ReceiptBranch::find($request->receipt_id);  
+            if($receipt->done){
+                toast('لا يمكن إضافة منتجات بعد الإنتهاء من الفاتورة','error');
+                return redirect()->back();
+            }
             $price_type = $receipt->price_type();
 
             $product = ReceiptBranchProduct::findOrFail($request->product_id);
@@ -449,6 +460,10 @@ class ReceiptBranchController extends Controller
 
     public function update(UpdateReceiptBranchRequest $request, ReceiptBranch $receiptBranch)
     {
+        if($receiptBranch->done){
+            toast('لا يمكن تعديل الفاتورة بعد الإنتهاء من الفاتورة','error');
+            return redirect()->back();
+        }
         $receiptBranch->update($request->all());
 
         toast(__('flash.global.update_title'),'success');
@@ -469,6 +484,10 @@ class ReceiptBranchController extends Controller
         abort_if(Gate::denies('receipt_branch_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $receiptBranch = ReceiptBranch::withTrashed()->find($id); 
+        if($receiptBranch->done){
+            toast('لا يمكن حذف الفاتورة بعد الإنتهاء من الفاتورة','error');
+            return redirect()->back();
+        }
         if($receiptBranch->deleted_at != null){
             $receiptBranch->forceDelete();
         }else{
